Rename Responder -> FormattedResponder
Preparation for adding a chained responder.
